Lift over a vector of genomic coordinates

Given a character vector of coordinates in the format "chrN:start-end", this function uses the UCSC liftOver utility (via the rtracklayer package) to map them to another assembly, for example from hg19 to hg38.

liftover_coordinates(
  coords,
  chain_file = "hg19ToHg38.over.chain",
  download_chain = FALSE,
  genome_style = TRUE
)

Arguments

coords

A character vector of coordinate strings like c("chr1:100-400", "chr3:450000-9890098").

chain_file

Path to the chain file for liftOver (e.g. "hg19ToHg38.over.chain"). If this file does not exist and download_chain=TRUE, it will be downloaded.

download_chain

Logical, if TRUE and chain_file does not exist, attempts to download the hg19->hg38 chain file from UCSC to the specified path.

genome_style

Logical, if TRUE, tries to maintain "chr" prefix. If your chain file expects a certain style, ensure your input coords have matching "chr" or no-"chr" style. Alternatively, you can remove or add the prefix before calling this function.

Value

A data frame with the following columns:

original

The original coordinate string.

mapped_chr

The chromosome (in new assembly).

mapped_start

The start position in new assembly.

mapped_end

The end position in new assembly.

status

Either "mapped" if a single unique mapping was found, "failed" if no mapping was found, or "split" if the range mapped to multiple segments.

Examples

if (FALSE) { # \dontrun{
liftover_coordinates(c("chr22:23632600-23632600", "chr9:133729451-133729600"),
                     chain_file="hg19ToHg38.over.chain")
} # }