Subset CpG Sites by GRanges Files

This function processes a vector of filenames containing CpG site data and subsets them based on overlaps with a provided GRanges object. It returns either a list of raw data.tables from each file or a metric matrix of (averaged) beta values.

subset_cpg_by_GR_files(filenames, gr, output = c("raw", "metric"))

Arguments

filenames

A character vector of filenames to be processed.

gr

A GRanges object. The names of gr (accessed via names(gr)) will be used to label the rows in the metric matrix.

output

A character string indicating the type of output to return. "raw" (the default) returns a list of raw data.tables with each table corresponding to a file. "metric" returns a matrix (or data.frame) where rows correspond to names(gr) and columns correspond to files, with cells containing the (averaged) beta values.

Value

If output = "raw", a list of data.table objects is returned, where each data.table has an additional column file indicating the source filename. If output = "metric", a matrix (or data.frame) is returned with rownames corresponding to names(gr) and columns corresponding to the files. For multiple overlaps in a given GRanges name, the beta values are averaged.

Details

For each file in filenames, the function reads in the data using fread() from the data.table package and then subsets the CpG sites that overlap with the GRanges object by calling subset_cpg_by_GR. Files yielding no overlaps are removed from the final results. For "metric" output, the function creates a vector (with NA for missing overlaps) for each file, computing the mean beta value for cases of multiple overlaps.

Examples

if (FALSE) { # \dontrun{
# Create an example GRanges object.
gr <- GRanges(seqnames = "chr1", ranges = IRanges(start = c(1, 101), width = 100),
              names = c("probe1", "probe2"))

# Example vector of file names.
file_vec <- c("file1.csv", "file2.csv")

# Get raw output (list of data.tables)
raw_output <- subset_cpg_by_GR_files(file_vec, gr, output = "raw")

# Get metric matrix output
metric_output <- subset_cpg_by_GR_files(file_vec, gr, output = "metric")
} # }